+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.1 under development
 
+- New #359: Add `GridView::keepColumnAttributesInEmptyCell()` to keep column body cell attributes on empty cells (@vjik)
 - Enh #358: Make dependency container in `GridView` constructor optional (@vjik)
 
 ## 1.2.0 September 02, 2026